+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ Changelog
 2.0.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Added an alias command ``/help`` for ``/commands``
+- Fixed usage alias commands overriding real commands
 
 
 2.0.5 (2018-11-26)